tweak tmux copy binding
This was "freezing" tmux's response to prefix commands (though terminals
inside tmux would still work fine, so I could `tmux detach/tmux attach`
and resolve the issue). Unclear why: something about running in the
foreground and maybe the process wasn't getting some ENV or something
and it hung? Running in a shell (and a bg one at that) seems to resolve
that problem.