move GNUPGHOME to pam_environment
That's allthe zsh file was doing now, this is more compact.

I played briefly with starting gpg-agent via systemd as well, but AFAIK the provided unit files are kinda broken: the socket unit specifies  as the dir for sockets, but newer gpg-agents but sockets under a semi-random subdir of that path (which isn't straightforwardly configurable as far as I can determine). So it's not clear to me how those units are intended to work given that behavior. Oh well, leave it alone ofor now.