Symlink contents of ~/.config, not .config directly
I go back and forth on this - symlinking the top-level dir is simple,
but kind of a hassle both for regular maintenance (I needed the big
.gitignore file), and also for setting up a new machine, which I'm doing
right now (because I had cause to create .config with some contents
before installing my dotfiles)

So for now I think I've come down on the side of... don't do that?
Keeping the config under `home/config` for now, although it makes the
installer logic a bit less consistent.