shell script to keep zoom up to date in aurto
Zoom's official packages are annoyingly provided this way. And they
don't mark a proper maintainer so aurto prompts to trust every time, so
I can't put this in a cron job (I suppose I could if I wrote an expect
script or maybe just did `echo y | aurto addpkg`?

Going to continue to not use the the aur package for now: relying on the
official package seems desirable to me, although the AUR package seems
to be almost entirely a "wrapper" for this at this point? But I'm not
making that decision just yet, so for now I'm adding this script so that
I can check for updates occassionally.