rm usb-wakeup stuff
I was playing with something like this since for some reason my system76
laptop was not by default getting woken up my the external keyboard.

It's been working fine more recently, and this script as-is wasn't
really functional anyway. Marking *every* USB device to wakeup is risky
and I found it made it impossible for the computer to actually stay
asleep at all. If I do need to revive this, I'll need to figure out how
to identify specific USB devices that should be touched.