Add firefox policies.json
This happens to be where FF is installed for me because I use the beta
channel. I think normally on Arch FF goes to /usr/lib/firefox.

Apparently, /etc/firefox/policies is a fallback path that can be used
regardless of the actual installation location, but I don't really want
to end up with an extra directory in /etc if I don't need to.

References:
- https://unixsheikh.com/articles/choose-your-browser-carefully.html#tweaking-firefox
- https://github.com/mozilla/policy-templates/blob/master/README.md