git: rm media config
I don't have a git-media binary installed, and my git history
doesn't really explain why I had this configured. I think the idea was
to automatically optimize pngs and svgs and things? Clearly not useful
right now since the binary isn't installed, so away it goes for now. Not
sure global config is really the best place for this anyway if I do use
it again in the future.