new-machine mkinitcpio - amend kernel opt for thinkpad screen freezes
The prior commit was a bad diagnosis since I was rushing and not reading
closely. That stack overflow post was related to hard freezes, which
isn't what I'm getting. What I'm seeing is the entire display becomes
unresponsive for a few seconds, then it's fine, and it only happens
using the laptop's own screen - not an external monitor.

I swore I hit this before, but I think it must have been on my old
ThinkPad, not my System76 machine, because I can't find any evidence of
a config for this on the System76 machine. Unfortunately I don't appear
to have documented the fix anywhere before, and my old ThinkPad's hard
drive has now been wiped, so I can't check. But I do recall that when I
debugged this last time it was related to "panel self refresh" (PSR), so
I think this is it. The Arch BBS post I found this time indicated a few
i915 parameters to flip, but I think PSR was all I needed last time, so
I'm going to try that for now.