Drop Cloudflare DNS - use Google instead
I've been noticing very Internet recently, and it appears to
be due to Cloudflare DNS.

Testing with `dig @1.1.1.1 google.com`, I would say ~30-50% of the time
the request times out, or takes several seconds to resolve. Querying via
8.8.8.8 is reliably fast. With 1.1.1.1 as my preferred DNS host in
resolv.conf, `curl https://www.google.com` often takes ~5s. With
8.8.8.8, it's reliably a few hundred ms. So I think what's happening is
that in many cases cloudflare is timing out, and then resolution falls
back to google, which then resolves the address quickly.

This has been going on for at least a couple weeks, I think, and I
finally got annoyed enough to try and figure out what was up (for a
while I just thought our ISP was worse than usual and it was a bandwidth
problem, before I noticed bandwidth-sensitive tasks like video were
fine, it was latency that was the problem).

I confess I'm very surprised at this diagnosis - it's clearly not some
kind of temporary incident, and if Cloudflare's public DNS was fubared
you'd think it might have generated some Internet chatter.  My best
theory right now is that some misconfiguration in a hop along the way
from me to 1.1.1.1 (probably at my ISP) is causing packets to get
mis-routed or dropped or something, so not many people are impacted (and
I'm living in a fairly sparse suburban area, so I might be literally the
only person in the area trying to use Cloudflare DNS).