redshift - override restart interval
redshift seems to sporadically crash out - seems like it might be when I wake my laptop up and it takes a moment for wifi to reconnect or something. when this happens redshift bounces quickly enough that it crashses too many times in a row and systemd stops trying to bring it up.

Trying the restart less often will hopefully fix this.