Switch backups from duplicity to restic
Restic is still a bit young, but of the "new" backup systems it's the
one that looks most stable & has B2 support.

Duplicity works, but for very large backup sets it has a pretty serious
flaw due to it's older-style architecture. Duplicity captures full
snapshots, and then "diff snapshots" on top of that snapshot,
effectively. Because of that, Duplicity can't do effective garbage
collection - it can't delete old unused files from a parent snapshot
even if those files have been gone from all the later snapshots for a
long time. And the longer you go without doing a clean snapshot, the
riskier a spot you're in for verifying your backups. So my impression
was the recommended usage for Duplicity was to do clean snapshots on a
semi-frequent schedule, with more frequent diff snapshots in between,
and regularly clean them up by deleting older snapshot sets. So I was
doing daily snapshots, with a new clean one each month.

All of that is fine if your clean snapshots are pretty quick, but my
clean backups with duplicity were taking over a day. That's acceptable
for, say, the very first setup of a backup system, but that's a hassle
to deal with *monthly*.

Restic doesn't work that way - it uploads blocks of data for files,
writes indexes to map file info to data, and then marks the state of
an index as being a snapshot. So after the initial backup, you can do
smaller backups indefinitely while also cleaning up old, unneeded
snapshot data.

Duplicity's other annoyance for me was that it's Python, and using B2 as
a backend required an additional python package. I do not think
scripting languages and their packaging systems tend to be a good fit
for system-type tools like this - they're too unstable. E.g. when I
installed a Python version upgrade, the package directory for `pip`
changed, and duplicity stopped working because the b2sdk package was no
longer installed in the right place & I had to install it again. Having
random package upgrades break your backup system isn't great! Restic is
a compiled go binary without external dependencies (aside from libc, I
presume).

I have opted not to rename some things - you'll note I used "duplicity"
in some of the password entry names originally, and I'm not going to
change that. I knew I was going to live to regret it when I did it, and
here we are. So I'll be making backups with restic to a B2 bucket named
for duplicity for the forseeable future.

Last little note: I still find systemd unit dependencies very confusing,
but I think I do want `Requires` in addition to `After` for
`network-online.target`? My understanding reading the man pages is
`After` controls ordering, but if the network *isn't* online, the unit
would still get run. Whereas doing both `After` & `Requires` basically
says "don't run this unit until the network is online, and if the
network isn't going to be online don't run it at all."