nfs mount - remove auto option
This option doesn't really do anything in this context, it would only
matter if I still had this in fstab. (And in fstab auto-mounting didn't
really work because at the time the system performs auto-mounts, the
network isn't up yet. Which is part of why I moved it to a systemd
unit!)