Don't load vim's netrw plugin
I've noticed vim taking unusually long to quit recently, and tracked it
down to this plugin.

I have a NAS at home for storage that I mount via NFS (there are mount &
automount systemd units for it in this repo). I'm away from home this
week and quitting vim was hanging for about a minute. Even at home I've
noticed vim having a short-but-noticeable (a second or two) lag when
quitting for a while, although I couldn't really pinpoint when it
started. (This lag occurred regardless of whether I was editing files on
the NFS mount - that should at least be obvious for the very long lag
this week since I'm not at home and cannot mount that NFS volume right
now.)

The long ~1 minute hang did seem like some kind of timeout, probably
network related, and some stack overflow posts suggested the same and
mentioned the netrw plugin as a likely culprit. I disabled the NFS
automount unit (so the systemd placeholder for the mount no longer
showed up in `mount`), and found that eliminated the lag exiting vim. I
then re-enabled the automount & added the change in this commit to
disable the netrw plugin, and that also eliminated the lag.

One of the first things I looked at was `/proc/<pid>/fd/` for the vim
process during the long exit hang, and that didn't show anything
interesting. So I'm not sure what netrw is trying to do on exit, but it
doesn't seem to be as straightforward as trying to read from mounts in
the normal way - or maybe it is but because this mount wasn't really
"mounted" and is interacting with systemd the fd was really a pipe of
some kind? Either way, I think it's interesting that the interaction was
the obvious answer of "it's trying to do someting with a networked mount
that isn't available", but that interaction did not show up in the
obvious way in `/proc`. Maybe `strace` would have revealed something
more.

I would like to figure out some kind of "location-based-systemd-target"
setup or something where I could activate some systemd units only when
I'm connected to my home WiFi network, but that seems complicated & I'm
not sure how to do it. Disabling the plugin seems more practical. Even
if I do figure out some weird systemd setup, I'd probably want to keep
the plugin disabled since the lag even when I was at home with the NFS
mounted was noticeable enough to be irritating. I've made it this far in
my life never once actually needing to use netrw's functionality, so I
don't expect to have any issues with this.