polybar - show status of some systemd units
Sometimes something will break in syncing imap and I don't really notice
until I see that mail isn't being propely synced to another device.

Mail & backups are kind of important - if they're failing, I want to
know. This adds a couple more polybar modules to show the current status
of the modules.

Another possibility I considered was a new timer unit that would run a
script to check all the other units I care about, and then notify me (by
mail or dunst) if any are failing. But that sort of runs into a
turtles-all-the-way-down problem of how do I know the unit that checks
the other units is healthy. Quis custodiet ipsos custodes and all that.
I could write that unit, and monitor *that* unit in a polybar check.

Anyway, I'm going to try this for now. If I need to monitor a bunch of
units, a bg process that mails me might be better, but with only 2 an
icon per service is fine. One nice perk of this approach is also that I
can see when the process in question is currently running.