polybar: show state of system update-aurto unit
I like aurto for managing AUR packages in a lot of ways, but it has its
rough spots. Since aurto works by building packages from the AUR in a
sandbox as packages that are then available in a local repository and
installed when you `pacman -Syu`, it regularly and automatically checks
for updates & builds updated packages.

Unfortunately this means that if a package fails to build it's not
really visible, you just won't see any updates for that package when you
`pacman -Syu`. Breakage isn't common but since the packages are built in
an isolated sandbox I've seen it happen due to dependency issues, etc.
Also, `update-aurto` goes through waiting packages in order and has no
failure handling, so if you install packages `A`, `B`, and `C` from the
AUR and `A` starts failing to build, you also won't see updates to
packages `B` or `C`.

Like other stuff, I'm not sure an ever-present icon in my menu bar is
exactly what I want, but it's easily available right now.