polybar - script to disable screen locker
It's convenient to have a quick way to disable the screen locker when
I'm, for example, on a video call. It's also convenient to have a UI
indicator so I don't forget to re-enable the locker after the call.

I was using caffeine-ng for a long time, but I found it surprisingly
finicky and unreliable. I realize that dealing with screen lockers on
Linux isn't exactly trivial since there's a bunch of them and they all
work a bit differently, but I don't think that's what was going on with
caffeine-ng: it wasn't so much "not working" as it was "usually working
but ocassionally hanging or not actually disabling the locker when it
thought it had". I've also learned to look askance at system tools
written in python if there's an alternative - there always seem to be
some packaging headaches associated with them and I've hit that a few
times with caffeine-ng as well.

So for a while I've just been running `xautolock -disable`/`-enable` in
a terminal as needed, which is fine, but when jumping on a call it's
nice to have it be a click in the bar & to have the visual reminder of
the current state there.