spideroak - try different Install.WantedBy
I noticed spideroak wasn't actually starting, though the unit was
enabled.

The noticeable differences of this unit from some others that run in my
systemd user daemon is this setting, and the `After` setting.

I'm hoping the `WantedBy` is the important difference, but I suspect
I'll also need to change the `After`. The user daemon doesn't *have*
`local-fs.target` or `network.target` - those only exist in the system
daemon.