runx: fix determination of XDG_RUNTIME_DIR
This was a simple case of not paying close enough attention and rushing.

The sed expression wasn't matching the null byte after the value, it was
matching the last zero in my uid (and dropping it from the output
string). So the determined XDG_RUNTIME_DIR was incorreect, and would be
even more incorrect for a user without a 0 in their pid (unusual as that
is these days).