nfs mount - soft option
The `hard` option is used by default. The trouble with the `hard` option
is that operations on the NFS volume are retried indefinitely and won't
fail (see `man nfs`). So if I have the NFS vol mounted and take my
laptop out of my home where the NFS vol is no longer accessible, some
things (like a running vim window) can get stuck and need to be
SIGKILL-ed.

I think `soft` is preferable for my use-case.