Run mkarchiso verbosely
I ran into a silent failure in a different dotfiles repo for this. A
package that is still in the default profile was apparently recently
removed from community and is now an AUR package, so it mkarchiso
couldn't find it. I just removed the package from the packages list, but
without this `-v` flag it can be hard to know what failed.