Add is-home-wifi script
Same basic logic as I needed for the udev script recently, but as its
own executable so I can use it for SSH Match rules. (Discussed in
http://will.flemi.ng/posts/2021/08/12/ssh-hostname-wifi/.)

This is a slight amount of duplication - I could have put this under
`system/...` as a bin and re-used it in the udev script, but for now
this is tolerable.