add explanatory comment to lidswitch.conf
This was in the copy copied to /etc - I must have added it directly
there at some point, but not here since this is one of the config files
that gets copied, not symlinked. Just came up since I added a new udev
rule file and my install script told me the files differed.