Add alias for generating passwords
For personal passwords I would of course use `pass generate`, but
sometimes I need to generate a password for some purpose that I don't
intend to keep in my password store. I generally end up on
https://www.howtogeek.com/howto/30184/10-ways-to-generate-a-random-password-from-the-command-line/
to remind myself of this invocation. (The `openssl` invocation is easier
to remember, but I'd rather not rely on having that installed.) I do it
often enough that I may as well have an alias.