sway: auto float/move Firefox sharing indicator window
When I start a video call the sharing indicator window is by default
treated as another tiling window, which tends to take over my whole
screen, and then I need to move back to the correct FF window. This
automates treating that window as floating and moving it out of the way.