new-machine: specify keyserver for aurutils author key
This was kind of weird. I switched to keys.ubuntu.com when the
sks-keyservers pool seemed to fall apart, and I'm sticking with it
because it's been pretty good, but on this *specific* key it was 500ing
yesterday. (And I could reproduce that from their web UI as well.) Every
other key I needed to pull from it worked fine, and I chose the ubuntu
keyserver because I've had some issues with the openpgp keyserver in the
past, but it seems to be a better place to pull this key from for now?