start gammastep & dropbox after sway starts
Following https://wiki.archlinux.org/title/Sway#Manage_Sway-specific_daemons_with_systemd

Dropbox was fine starting first, it just meant the indicator didn't
appear in the systray. It's nice to have that, so I moved it here.

Gammastep has to start after sway. By default, this would be enabled for
graphical-session.target. That target was not starting for me,
presumably because I invoke sway directly from a tty. Now that I'm
triggering sway-session.target which also starts
graphical-session.target, I could just use gammastep's default
`[Install]`, but this is more specific so I'll stick with it (in case I
ever need to dual-run an X11 env again it'll be helpful).