vim: tweak GFiles invocation
This basically just runs git-ls-files under the hood and shows the
results with fzf. Default behavior is just `--cached`, but it's useful
to also see new files I'm working on but haven't `git added` yet. That's
`--others`, but then that also means listing `gitignored` files, so
`--exclude-standard` excludes those again.