Switch restic to use the s3 api for backblaze
Restic's [docs][1] recommend this over the native B2 API for error
handling reasons.

The only tweak needed to the config stored in pass entries is the bucket
needs to include the backblaze s3 api endpoint domain (which is listed
with the bucket name in backblaze's console). I.e. instead of storing
"foo", it's like "s3.us-<region>-<nnn>.backblaze2.com".

[1]: https://restic.readthedocs.io/en/latest/030_preparing_a_new_repo.html#backblaze-b2