Add slack wrapper script
I'm about to immediately remove this, I'm adding it briefly as
documentation. I've had it locally for a while without committing it to
resolve an issue I was having where slack would launch with a blank
window, which seemed to be a bug related to Electron & my GPU driver.
That issue seems resolved now, so I'm removing the wrapper but adding it
in this commit very briefly in case it happens again so I can recover it
quickly.