Tweak backup service to retry
Since my laptop is often asleep when the timer for this triggers, it
doesn't actually start until I wake up my laptop in the morning, and
then frequently seems to immediately fail since the network isn't
actually connected yet.

I would like to be able to `Requires=network-online.target` here (I'm
still not very confident of my understanding of some systemd directives
like this, but I think `After` only affects *ordering*, while `Requires`
adds a dependency on *state*), but don't think I can. `After` seems to
be ok for a user unit referencing a system unit, but not `Requires`: it
errors about the target not existing.  The error for `Requires` makes we
wonder if `After` is also just silently ignoring these targets and not
actually doing anything, but let's put that aside for now.

If I can't properly wait for the network to be online this way, I
can at least add some retry logic so the backup should hopefully retry
after the network is connected and then succeed.

An ExecStartPre using netcat to try cloudflare's reliable DNS address is
also a good way to make the initial start a bit more reliable by trying
to wait for the network to be online.