slack binstub
I've had this locally for a while and never committed it since it was
meant to be temporary. I forget which option here exactly was needed or
what the symptom was. I think it was a crash and the platform hint was
the important bit here? Either way, doesn't seem necessary anymore, so
I'm about to delete this but I'm committing it for a moment as
documentation of what I needed for a while in case something similar
recurs.