fix some nvim settings
I thought these options could be treated as tables, but that appears to
be incorrect. `:h vim.opt` does suggest `:append` as an alternative to
old vimscript `set ... +=`, so I think these are class instances (since
`:foo` is related to class/methods in Lua as I understand it.) And I
guess `table.insert` isn't an error because Lua objects are just fancy
structs/tables.