fix swp file directory
Weirdly the handling of ~ in this setting seems to depend on how it's
set. This was what I had in my init.vim for years and it worked, but in
lua this results in a directory literally named '~' getting created in
the current directory.

I was going to fix by using `vm.fn.expand("$HOME/...")`, but while
checking what the behavior was in a vimrc config by starting a vim
session without an rc file (`vim -u NORC`), I discoevered that
apparently had some point nvim changed the default swap directory to a
location under `~/.local/state/nvim`, which is a fine default, so
bye-bye silly config I've had to work around a braindead default that's
been in original vim for probably a few decades.