Add git-fix-file helper script
On larger feature branches I'm often working on multiple files in
sequence and ammending previous commits as changes come up. I manually
lookup the previous commits, which is kind of a hassle and also
error-prone. E.g. I might be changing a model schema, writing the
migration, then updating the backend, then updating the frontend, and
might realize I need to adjust backend logic as I'm changing the
frontend or iterate on the schema.

This automates a lot of that drudgery by looking up the last commit a
file was changed in and generating a fixup commit for the file pointing
at the correct recent parent.