change mutt mark-as-spam binding
Had originally chosen "s" for "spam" but that's right next to "a" for
"archive", and also the same key as standard mutt binding for "save". So
I'm constantly hesitating when I mark something as spam or save to a
folder to double check I didn't do the wrong thing. I'm gonna switch to
"j" for "junk".